requirements.txt
asgiref==3.2.10
Django==3.1
django-cors-headers==3.10.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
environ==1.0
factory-boy==3.2.1
Faker==13.13.0
flake8==4.0.1
mccabe==0.6.1
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
